pyproject.toml
[project]
name = "cyberthonmlops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.136.0",
    "uvicorn>=0.45.0",
    "asyncpg>=0.31.0",
    "boto3>=1.42.87",
    "boto3-stubs~=1.42.87",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.12",
]
